ai-services/requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
mysql-connector-python>=8.0.0

# Data Processing & ML (minimal versions)
pandas>=2.0.0
numpy>=1.20.0

# Basic ML (will add more complex ones later)
scikit-learn>=1.0.0

# For now, we'll implement basic versions without heavy NLP libraries
# sentence-transformers
# transformers
# torch

# HTTP Client
httpx>=0.25.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
structlog>=23.0.0

# Optional: For advanced analytics
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.17.0

# Optional: For vector databases (if scaling up)
# chromadb==0.4.18
# pinecone-client==2.2.4

# Performance optimization
# numba removed - not compatible with Python 3.13
joblib>=1.3.0
